# reference/apisperu-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "API REST Facturaci\u00f3n Electr\u00f3nica - APIsPERU. <br /> Si se siente m\u00e1s c\u00f3modo usando Swagger-UI, puede <a target='_blank' href='/doc/swagger/ui'>ingresar aqu\u00ed.</a>",
        "version": "1.3",
        "title": "API DE FACTURACI\u00d3N",
        "contact": {
            "name": "Apisperu",
            "email": "soporte@apisperu.com"
        },
        "x-logo": {
            "url": "/img/logoredoc.png",
            "altText": "Logo APIsPERU"
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/companies": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Obtiene el listado de empresas",
                "description": "Retorna la lista de empresas del usuario actual",
                "operationId": "company_index",
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            },
            "post": {
                "tags": [
                    "company"
                ],
                "summary": "Crear un nueva empresa",
                "description": "Crea una empresa nueva para el usuario actual.",
                "operationId": "company_new",
                "requestBody": {
                    "description": "Datos de la empresa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "plan": {
                                                "description": "Utilice una de estas opciones: 'free', 'premium'.",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "environment": {
                                                "description": "Utilice una de estas opciones: 'beta', 'produccion', 'nubefact_beta' o 'nubefact_produccion'. <br> <b>beta</b>: utiliza los ambientes de prueba de la Sunat. <br> <b>produccion</b>: utiliza los ambientes de producci\u00f3n de la Sunat. <br> <b>nubefact_beta</b>: utiliza los ambientes de prueba de nubefact. <br> <b>nubefact_produccion</b>: utiliza los ambientes de producci\u00f3n de nubefact",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/MyCompany"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
        "/companies/{companyId}": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Retorna una empresa",
                "description": "Obtiene una empresa en particular del usuario actual",
                "operationId": "company_show",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "required": true,
                        "description": "Id de la empresa",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "description": "",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "plan": {
                                                    "description": "",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "environment": {
                                                    "description": "",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/MyCompany"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "company"
                ],
                "summary": "Editar una empresa",
                "description": "Actualiza una empresa determinada",
                "operationId": "company_edit",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "required": true,
                        "description": "Id de la empresa",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos de la empresa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "plan": {
                                                "description": "Utilice una de estas opciones: 'free', 'premium'.",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "environment": {
                                                "description": "Utilice una de estas opciones: 'beta', 'produccion'. <br> beta: utiliza los ambientes de prueba de la Sunat. <br> produccion: utiliza los ambientes de producci\u00f3n de la Sunat.",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/MyCompany"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operacion exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "description": "",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "plan": {
                                                    "description": "",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "environment": {
                                                    "description": "",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/MyCompany"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "company"
                ],
                "summary": "Eliminar una empresa",
                "description": "Elimina una empresa determinada",
                "operationId": "company_delete",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "required": true,
                        "description": "Id de la empresa",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "description": "",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "plan": {
                                                    "description": "",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "environment": {
                                                    "description": "",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/MyCompany"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/companies/certificate/free": {
            "post": {
                "tags": [
                    "company"
                ],
                "summary": "Genera un certificado PFX gratuito para hacer pruebas",
                "description": "Esta ruta genera un certificado PFX gratuito, Usar solo para pruebas.",
                "operationId": "company_certificate_free",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "description": "La contrase\u00f1a para proteger el certificado PFX"
                                    },
                                    "ruc": {
                                        "type": "string",
                                        "description": "El RUC de la empresa."
                                    }
                                },
                                "required": [
                                    "password",
                                    "ruc"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Archivo PFX generado exitosamente",
                        "content": {
                            "application/x-pkcs12": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud, posiblemente debido a una contrase\u00f1a faltante o archivo demasiado grande"
                    },
                    "500": {
                        "description": "Error interno del servidor al generar el archivo PFX"
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "https://facturacion.apisperu.com/api/v1"
        }
    ],
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Value: Bearer {jwt}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        },
        "schemas": {
            "MyCompany": {
                "required": [
                    "ruc",
                    "razon_social",
                    "direccion",
                    "certificado",
                    "logo",
                    "sol_user",
                    "sol_pass",
                    "plan",
                    "environment"
                ],
                "properties": {
                    "sol_user": {
                        "description": "Usuario secundario SUNAT con permisos para emisi\u00f3n electr\u00f3nica. <br />Siga este tutorial para crear un usuario secundario SUNAT: <a href='https://youtu.be/9Q9rSKSL7Os' target='_blank'>https://youtu.be/9Q9rSKSL7Os</a>",
                        "type": "string"
                    },
                    "sol_pass": {
                        "description": "Contrase\u00f1a usuario secundario SUNAT. <br />Siga este tutorial para crear un usuario secundario SUNAT: <a href='https://youtu.be/9Q9rSKSL7Os' target='_blank'>https://youtu.be/9Q9rSKSL7Os</a>",
                        "type": "string"
                    },
                    "ruc": {
                        "type": "string"
                    },
                    "razon_social": {
                        "type": "string"
                    },
                    "direccion": {
                        "type": "string"
                    },
                    "certificado": {
                        "description": "Certificado PEM en formato Base64 String. <br />Puede usar la api <a href='/doc#operation/company_cert'>Convertir P12, PFX </a> para generar el archivo PEM en Base64. </a> Tama\u00f1o m\u00e1ximo permitido 200KB. <br /><br /> Si desea realizar pruebas puede generar un certificado gratuito desde el endpoint <a href='/doc#tag/company/operation/company_certificate_free'>/companies/certificate/free</a>",
                        "type": "string"
                    },
                    "logo": {
                        "description": "Logo png en formato Base64 String. Tama\u00f1o m\u00e1ximo permitido 200KB.",
                        "type": "string"
                    },
                    "client_id": {
                        "description": "Credencial generado en el portal de SUNAT. Revisar este tutorial: <a href='https://youtu.be/y_-qfGpYKGA?si=fiVfvdq8FVDXW-M9' target='_blank'>https://youtu.be/y_-qfGpYKGA?si=fiVfvdq8FVDXW-M9 </a>. Opcional si desea usar la API SUNAT (requerido para la nueva versi\u00f3n de gu\u00edas de remisi\u00f3n)",
                        "type": "string"
                    },
                    "client_secret": {
                        "description": "Credencial generado en el portal de SUNAT. Revisar este tutorial: <a href='https://youtu.be/y_-qfGpYKGA?si=fiVfvdq8FVDXW-M9' target='_blank'>https://youtu.be/y_-qfGpYKGA?si=fiVfvdq8FVDXW-M9 </a>. Opcional si desea usar la API SUNAT (requerido para la nueva versi\u00f3n de gu\u00edas de remisi\u00f3n)",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}